requirements.txt
dnspython
gevent
requests
beautifulsoup4
pymongo
pyOpenSSL
ldap3
mysqlclient
mysql-connector
pyCrypto
pycryptodome
netaddr
paramiko
tqdm
elasticsearch
impacket
pillow
psycopg2-binary
redis
pypsrp
chardet
gevent
